+++ packages/dm-core-plugins/src/form/templates/ObjectModelContainedTemplate.tsx
@@ -1,4 +1,4 @@
-import { TObjectTemplate } from '../types'
+import { TFormConfig, TObjectTemplate } from '../types'
 import React, { useState } from 'react'
 import { useFormContext } from 'react-hook-form'
 import { useRegistryContext } from '../context/RegistryContext'
@@ -7,11 +7,12 @@ import AddObject from '../components/AddObjectButton'
 import { OpenObjectButton } from '../components/OpenObjectButton'
 import { ViewCreator } from '@development-framework/dm-core'
 import AddObjectBySearchButton from '../components/AddObjectBySearchButton'
-import FormObjectBorder from './shared/FormObjectBorder'
-import ObjectLegendHeader from './shared/ObjectLegendHeader'
-import FormExpandedViewWrapper from './shared/FormExpandedViewWrapper'
-import ObjectLegendWrapper from './shared/ObjectLegendWrapper'
-import ObjectLegendActionsWrapper from './shared/ObjectLegendActionsWrapper'
+import FormTemplate from './shared/FormTemplate'
+import {
+  getExpandViewConfig,
+  getOpenViewConfig,
+  getCanOpenOrExpand as getCanOpenExpand,
+} from './shared/utils'
 
 export const ObjectModelContainedTemplate = (
   props: TObjectTemplate
@@ -26,43 +27,39 @@ export const ObjectModelContainedTemplate = (
   const value = watch(namePath)
   const objectIsNotEmpty = value && Object.keys(value).length > 0
 
-  const canOpenInTab =
-    objectIsNotEmpty &&
-    onOpen &&
-    (uiAttribute?.functionality?.open ?? config.functionality.open)
-
-  const canExpand =
-    objectIsNotEmpty &&
-    (uiAttribute?.functionality?.expand ?? config.functionality.expand)
-
-  const openInTabViewConfig = uiAttribute?.openViewConfig
-    ? uiAttribute?.openViewConfig
-    : {
-        type: 'ReferenceViewConfig',
-        scope: namePath,
-        recipe: uiAttribute?.uiRecipe,
-      }
+  const { canExpand, canOpen } = getCanOpenExpand(
+    objectIsNotEmpty,
+    config,
+    uiAttribute,
+    onOpen
+  )
   return (
-    <FormObjectBorder>
-      <ObjectLegendWrapper>
-        <ObjectLegendHeader
+    <FormTemplate>
+      <FormTemplate.Header>
+        <FormTemplate.Header.Title
           canExpand={canExpand}
-          canOpenInTab={canOpenInTab}
+          canOpen={canOpen}
           isExpanded={isExpanded}
           attribute={attribute}
           objectIsNotEmpty={objectIsNotEmpty}
           setIsExpanded={setIsExpanded}
-          openInTab={() => onOpen?.(namePath, openInTabViewConfig, idReference)}
+          onOpen={() =>
+            onOpen?.(
+              namePath,
+              getOpenViewConfig(uiAttribute, namePath),
+              idReference
+            )
+          }
         />
-        <ObjectLegendActionsWrapper>
-          {canOpenInTab && (
+        <FormTemplate.Header.Actions>
+          {canOpen && (
             <OpenObjectButton
               viewId={namePath}
               idReference={idReference}
-              viewConfig={openInTabViewConfig}
+              viewConfig={getOpenViewConfig(uiAttribute, namePath)}
             />
           )}
-          {attribute.optional && !config.readOnly && (
+          {attribute.optional && !config.readOnly && !objectIsNotEmpty && (
             <>
               {uiAttribute?.searchByType && (
                 <AddObjectBySearchButton
@@ -86,25 +83,18 @@ export const ObjectModelContainedTemplate = (
               namePath={namePath}
             />
           )}
-        </ObjectLegendActionsWrapper>
-      </ObjectLegendWrapper>
+        </FormTemplate.Header.Actions>
+      </FormTemplate.Header>
       {canExpand && isExpanded && (
-        <FormExpandedViewWrapper>
+        <FormTemplate.Content>
           <ViewCreator
             idReference={`${idReference}.${namePath}`}
             onOpen={onOpen}
-            viewConfig={
-              uiAttribute?.expandViewConfig
-                ? uiAttribute?.expandViewConfig
-                : {
-                    type: 'ReferenceViewConfig',
-                    recipe: uiAttribute?.uiRecipe,
-                  }
-            }
+            viewConfig={getExpandViewConfig(uiAttribute)}
             onChange={(data: any) => setValue(namePath, data)}
           />
-        </FormExpandedViewWrapper>
+        </FormTemplate.Content>
       )}
-    </FormObjectBorder>
+    </FormTemplate>
   )
 }