+++ packages/dm-core-plugins/src/form/templates/ObjectStorageUncontainedTemplate.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { useFormContext } from 'react-hook-form'
 import {
   resolveRelativeAddress,
@@ -7,18 +7,22 @@ import {
 } from '@development-framework/dm-core'
 import { useRegistryContext } from '../context/RegistryContext'
 import { TObjectTemplate } from '../types'
-import { Fieldset, Legend } from '../styles'
-import { Typography } from '@equinor/eds-core-react'
-import { getDisplayLabel } from '../utils/getDisplayLabel'
 import RemoveObject from '../components/RemoveObjectButton'
 import { OpenObjectButton } from '../components/OpenObjectButton'
 import { SelectReference } from '../components/SelectReference'
+import FormTemplate from './shared/FormTemplate'
+import {
+  getCanOpenOrExpand,
+  getExpandViewConfig,
+  getOpenViewConfig,
+} from './shared/utils'
 
 export const ObjectStorageUncontainedTemplate = (props: TObjectTemplate) => {
   const { namePath, attribute, uiAttribute } = props
   const { watch, setValue } = useFormContext()
   const { idReference, onOpen } = useRegistryContext()
   const { dataSource, documentPath } = splitAddress(idReference)
+  const [isExpanded, setIsExpanded] = useState(false)
   const { config } = useRegistryContext()
   const value = watch(namePath)
   const address = resolveRelativeAddress(
@@ -28,59 +32,56 @@ export const ObjectStorageUncontainedTemplate = (props: TObjectTemplate) => {
   )
 
   const referenceExists = address !== undefined
-  const canExpand =
-    referenceExists &&
-    (!onOpen ||
-      (uiAttribute?.functionality?.expand ?? config.functionality.expand))
-  const canOpen =
-    referenceExists &&
-    onOpen &&
-    (uiAttribute?.functionality?.open ?? config.functionality.open)
+
+  const { canExpand, canOpen } = getCanOpenOrExpand(
+    referenceExists,
+    config,
+    uiAttribute
+  )
 
   return (
-    <Fieldset>
-      <Legend>
-        <Typography bold={true}>{getDisplayLabel(attribute)}</Typography>
-        {!config.readOnly && (
-          <SelectReference
-            attributeType={attribute.attributeType}
-            namePath={namePath}
-          />
-        )}
-        {attribute.optional && address && !config.readOnly && (
-          <RemoveObject address={address} namePath={namePath} />
-        )}
-        {canOpen && (
-          <OpenObjectButton
-            viewId={namePath}
-            viewConfig={
-              uiAttribute?.openViewConfig
-                ? uiAttribute?.openViewConfig
-                : {
-                    type: 'ReferenceViewConfig',
-                    scope: namePath,
-                    recipe: uiAttribute?.uiRecipe,
-                  }
-            }
-            idReference={address}
-          />
-        )}
-      </Legend>
-      {canExpand && (
-        <ViewCreator
-          idReference={address}
-          onOpen={onOpen}
-          viewConfig={
-            uiAttribute?.expandViewConfig
-              ? uiAttribute?.expandViewConfig
-              : {
-                  type: 'ReferenceViewConfig',
-                  recipe: uiAttribute?.uiRecipe,
-                }
+    <FormTemplate>
+      <FormTemplate.Header>
+        <FormTemplate.Header.Title
+          canExpand={canExpand}
+          canOpen={canOpen}
+          isExpanded={isExpanded}
+          attribute={attribute}
+          objectIsNotEmpty={referenceExists}
+          setIsExpanded={setIsExpanded}
+          onOpen={() =>
+            onOpen?.(namePath, getOpenViewConfig(uiAttribute), address)
           }
-          onChange={(data: any) => setValue(namePath, data)}
         />
+        <FormTemplate.Header.Actions>
+          {canOpen && referenceExists && (
+            <OpenObjectButton
+              viewId={namePath}
+              viewConfig={getOpenViewConfig(uiAttribute)}
+              idReference={address}
+            />
+          )}
+          {!config.readOnly && (
+            <SelectReference
+              attributeType={attribute.attributeType}
+              namePath={namePath}
+            />
+          )}
+          {attribute.optional && referenceExists && !config.readOnly && (
+            <RemoveObject address={address} namePath={namePath} />
+          )}
+        </FormTemplate.Header.Actions>
+      </FormTemplate.Header>
+      {canExpand && referenceExists && isExpanded && (
+        <FormTemplate.Content>
+          <ViewCreator
+            idReference={address}
+            onOpen={onOpen}
+            viewConfig={getExpandViewConfig(uiAttribute)}
+            onChange={(data: any) => setValue(namePath, data)}
+          />
+        </FormTemplate.Content>
       )}
-    </Fieldset>
+    </FormTemplate>
   )
 }